Update mongodb version

+++ src/index.js
@@ -71,30 +71,34 @@ async function mainLoop(cfg, client, slack) {
         reportLongRunningOps(slack, ops);
         if (ops.length !== 0) {
             _.forEach(ops, async (op) => {
-                console.log(`Detected a long running operation with ID: ${op.opid}`);
-                if (cfg.recordAllLongOps) {
-                    let res = recordOp(client, op, cfg.longOpsDB, cfg.longOpsCollection)
-                        .catch(() => { });
-                    promises.push(res);
-                }
-                if (cfg.killingEnabled) {
-                    if (op.secs_running > cfg.killThresholdSeconds) {
-                        if (matchesFilter(op, cfg.killFilter)) {
-                            let res = killOp(client, op)
-                                .then(() => {
-                                    console.log("Sending notification about the killed operation to Slack...");
-                                    return slack.alert('Killed a long running operation', op);
-                                })
-                                .catch((e) => {
-                                    return slack.alert(`Failed killing an operation, error: ${e}`, op);
-                                });
-                            _.remove(ops, (o) => { return o.opid == op.opid });
-                            return promises.push(res);
+                if (!op) {
+                    console.log(`Detect op that is unknown, skipping`);
+                } else {
+                    console.log(`Detected a long running operation with ID: ${op.opid}`);
+                    if (cfg.recordAllLongOps) {
+                        let res = recordOp(client, op, cfg.longOpsDB, cfg.longOpsCollection)
+                            .catch(() => { });
+                        promises.push(res);
+                    }
+                    if (cfg.killingEnabled) {
+                        if (op.secs_running > cfg.killThresholdSeconds) {
+                            if (matchesFilter(op, cfg.killFilter)) {
+                                let res = killOp(client, op)
+                                    .then(() => {
+                                        console.log("Sending notification about the killed operation to Slack...");
+                                        return slack.alert('Killed a long running operation', op);
+                                    })
+                                    .catch((e) => {
+                                        return slack.alert(`Failed killing an operation, error: ${e}`, op);
+                                    });
+                                _.remove(ops, (o) => { return o.opid == op.opid });
+                                return promises.push(res);
+                            }
+                            console.log(`Operation ${op.opid} doesn't match the kill filter, ignoring`);
+                            return slack.alert(`Detected an operation above the kill threshold, but not matching the kill filter`, op);
+                        } else {
+                            console.log(`Operation ${op.opid} is running less than ${cfg.killThresholdSeconds} seconds, delaying killing`);
                         }
-                        console.log(`Operation ${op.opid} doesn't match the kill filter, ignoring`);
-                        return slack.alert(`Detected an operation above the kill threshold, but not matching the kill filter`, op);
-                    } else {
-                        console.log(`Operation ${op.opid} is running less than ${cfg.killThresholdSeconds} seconds, delaying killing`);
                     }
                 }
